Break factory-agent.ts into smaller files
Split the monolithic factory-agent.ts into three focused modules:

- factory-agent-types.ts: shared types (ProjectCard, TicketCard,
  AgentContext, TestResult, etc.), constants, and resolveFactoryModel()
- factory-agent-tool-use.ts: ToolUseFactoryAgent (native tool-use
  protocol) and MockLoopAgent
- factory-agent.ts: OpenRouterFactoryAgent (declarative plan() model),
  validation functions, MockFactoryAgent, and barrel re-exports

Move the tool-use system prompt from inline construction to a template
file at prompts/system-tool-use.md. The old prompts (system.md,
action-schema.md, ticket-iterate.md) remain for the declarative agent.

All existing imports from './factory-agent' continue to work via
re-exports.
